improv new lint rule codegen
