+++ crates/rome_js_analyze/src/semantic_analyzers/nursery/use_hook_at_top_level.rs
@@ -9,7 +9,10 @@ use rome_rowan::AstNode;
 use super::use_exhaustive_dependencies::ReactExtensiveDependenciesOptions;
 
 declare_rule! {
-    /// <DESCRIPTION>
+    /// Enforce that all React hooks are being called from the Top Level
+    /// component functions.
+    /// 
+    /// To understand why this required see https://reactjs.org/docs/hooks-rules.html#only-call-hooks-at-the-top-level
     ///
     /// ## Examples
     ///
