+++ website/src/pages/lint/rules/index.mdx
@@ -656,6 +656,13 @@ Enforce all dependencies are correctly specified.
 Disallow the use of <code>Math.pow</code> in favor of the <code>**</code> operator.
 </section>
 <section class="rule">
+<h3 data-toc-exclude id="useHookAtTopLevel">
+	<a href="/lint/rules/useHookAtTopLevel">useHookAtTopLevel</a>
+</h3>
+Enforce that all React hooks are being called from the Top Level
+component functions.
+</section>
+<section class="rule">
 <h3 data-toc-exclude id="useNumericLiterals">
 	<a href="/lint/rules/useNumericLiterals">useNumericLiterals</a>
 </h3>
